归档/flip_camera_android/requirements.txt
shoots
shoots-android
shoots-uibase
opencv-python~=4.3.0.36
pathlib2~=2.3.5

pandas~=1.0.5
